GPUBurnerTask: Set get_time_spend_milliseconds as long long

+++ gpu_burner/GPUBurnerTask.cpp
@@ -13,6 +13,6 @@ void GPUBurnerTask::burn() {
     _end_milliseconds = duration_cast<milliseconds>(system_clock::now().time_since_epoch());
 }
 
-int GPUBurnerTask::get_time_spend_milliseconds() {
-    return _end_milliseconds.count() - _start_milliseconds.count();
+long long GPUBurnerTask::get_time_spend_milliseconds() {
+    return  _end_milliseconds.count() - _start_milliseconds.count();
 }
