+++ gpu_burner/GPUBurnerTask.h
@@ -40,7 +40,7 @@ public:
 
     virtual void recycle() = 0;
 
-    int get_time_spend_milliseconds();
+    long long get_time_spend_milliseconds();
 
 };
